+++ src/real/constants.mjs
@@ -1,12 +1,15 @@
 // src/real/constants.mjs
 export const OWNER_ALLOWLIST = [
-    'younestsouli2019@gmail.com', // PayPal
-    '007810000448500030594182',   // Bank Wire
+    'younestsouli2019@gmail.com', // PayPal/Payoneer Primary
+    '007810000448500030594182',   // Bank Wire (Attijari) - Also target for PayPal/Stripe
     'younesdgc@gmail.com',        // Secondary/Legacy
     '15924956',
     '231486',
     '4671926',
     'LU774080000041265646',
     'Barclays:231486:15924956',
-    'MUFG:0005:869:4671926'
+    'MUFG:0005:869:4671926',
+    '0xA46225a984E2B2B5E5082E52AE8d8915A09fEfe7', // Trust Wallet
+    '0xf6b9e2fcf43d41c778cba2bf46325cd201cc1a10', // Bybit ERC20
+    'UQDIrlJp7NmV-5mief8eNB0b0sYGO0L62Vu7oGX49UXtqlDQ' // Bybit TON
 ];